requirements.txt
salmon-mail==2
django==1.7
#psycopg2==2.3.2
django-forms-bootstrap
pinax-theme-bootstrap
#pinax-theme-bootstrap-account
django-user-accounts
django-bitfield
